fix:clubCategoryModalCSS
